demo/src/test/java/com/example/ApplicationTest.java
package com.example;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvFileSource;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;

class ApplicationTest {

    private final PrintStream printStream = System.out;

    private final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();

    @BeforeEach
    public void setUp(){
        System.setOut(new PrintStream(outputStream));
    }

    @AfterEach
    public void tearDown(){
        System.setOut(printStream);
    }


    @ParameterizedTest
    @CsvFileSource(resources = "/csv/notas.csv")
    void itShouldReturnTheCorrectLetterForTheCalification(int calification, String letra) {
        assertEquals(letra, Application.obtenerNota(calification));
    }
}
